+++ base/cvd/cuttlefish/host/commands/cvd/fetch/credential_flags.h
@@ -1,0 +1,37 @@
+//
+// Copyright (C) 2023 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <string>
+#include <vector>
+
+#include "cuttlefish/common/libs/utils/flag_parser.h"
+
+namespace cuttlefish {
+
+inline constexpr bool kDefaultUseGceMetadata = false;
+inline constexpr char kDefaultCredentialFilepath[] = "";
+inline constexpr char kDefaultServiceAccountFilepath[] = "";
+
+struct CredentialFlags {
+  std::vector<Flag> Flags();
+
+  bool use_gce_metadata = kDefaultUseGceMetadata;
+  std::string credential_filepath = kDefaultCredentialFilepath;
+  std::string service_account_filepath = kDefaultServiceAccountFilepath;
+};
+
+}  // namespace cuttlefish
