+++ base/cvd/cuttlefish/host/commands/cvd/fetch/fetch_cvd_parser.h
@@ -25,6 +25,7 @@
 
 #include "cuttlefish/common/libs/utils/result.h"
 #include "cuttlefish/host/commands/cvd/cache/cache.h"
+#include "cuttlefish/host/commands/cvd/fetch/credential_flags.h"
 #include "cuttlefish/host/libs/web/android_build_api.h"
 #include "cuttlefish/host/libs/web/android_build_string.h"
 #include "cuttlefish/host/libs/web/cas/cas_flags.h"
@@ -32,9 +33,6 @@
 
 namespace cuttlefish {
 
-inline constexpr bool kDefaultUseGceMetadata = false;
-inline constexpr char kDefaultCredentialFilepath[] = "";
-inline constexpr char kDefaultServiceAccountFilepath[] = "";
 inline constexpr char kDefaultApiKey[] = "";
 inline constexpr char kDefaultCredentialSource[] = "";
 inline constexpr char kDefaultProjectID[] = "";
@@ -49,12 +47,6 @@ inline constexpr bool kDefaultKeepDownloadedArchives = false;
 
 inline constexpr char kDefaultBuildTarget[] =
     "aosp_cf_x86_64_only_phone-userdebug";
-
-struct CredentialFlags {
-  bool use_gce_metadata = kDefaultUseGceMetadata;
-  std::string credential_filepath = kDefaultCredentialFilepath;
-  std::string service_account_filepath = kDefaultServiceAccountFilepath;
-};
 
 struct BuildApiFlags {
   std::string api_key = kDefaultApiKey;
